+++ test/hf/depstar_test.clj
@@ -152,6 +152,33 @@
       (is (deref #'sut/multi-release?))
       (is (= "true" (get-in contents [:manifest "Multi-Release"]))))))
 
+(deftest issue-64-jvm-opts
+  (println "[#64/JVM-OPTS]")
+  (let [jar (File/createTempFile "test" ".jar")
+        res (atom nil)]
+    (is (re-find
+         #"We are direct linking!"
+         (with-out-str
+           (reset! res
+                   (sut/build-jar {:jar-type :uber :jar (str jar)
+                                   :aot true :main-class 'issue-64
+                                   :jvm-opts ["-Dclojure.compiler.direct-linking=true"]
+                                   :pom-file (str (File/createTempFile "pom" ".xml"))
+                                   :group-id "issue" :artifact-id "bug" :version "64"
+                                   :aliases [:test-issue-64]})))))
+    (is (= {:success true} @res))
+    (is (not (re-find
+              #"We are direct linking!"
+              (with-out-str
+                (reset! res
+                        (sut/build-jar {:jar-type :uber :jar (str jar)
+                                        :aot true :main-class 'issue-64
+                                        ;:jvm-opts ["-Dclojure.compiler.direct-linking=true"]
+                                        :pom-file (str (File/createTempFile "pom" ".xml"))
+                                        :group-id "issue" :artifact-id "bug" :version "64"
+                                        :aliases [:test-issue-64]}))))))
+    (is (= {:success true} @res))))
+
 (deftest issue-66
   (println "[#66]")
   (let [jar (File/createTempFile "test" ".jar")]
@@ -159,8 +186,4 @@
            (sut/build-jar {:jar-type :uber :jar (str jar)
                            :pom-file (str (File/createTempFile "pom" ".xml"))
                            :group-id "issue" :artifact-id "bug" :version "66"
-                           :aliases [:test-issue-66]})))
-    #_(let [contents (read-jar jar #"module-info.class")]
-      ;; verify module-info.class not present:
-        (is (not (some #(= "module-info.class" %) (:entries contents))))
-        (is (empty? (:files contents))))))
+                           :aliases [:test-issue-66]})))))